DAO de gift creat; falta millorar el codi pero ja ho faré demà
